No idea what I did there...

+++ install.sh
@@ -5,6 +5,8 @@ mkdir -p ~/dotfiles_old
 cd ~/dotfiles
 
 for FILE in $FILES; do
-	mv ~/.$FILE ~/dotfiles_old
+	if [ -e ~/.$FILE ]; then
+		mv ~/.$FILE ~/dotfiles_old
+	fi
 	ln -s ~/dotfiles/$FILE ~/.$FILE
 done
